refactor(MapPolyfill): move polyfill definition to its own module for testing
